refactor(capabilities): migrate the component host cap to the struct-hosted actor macro
Move the ComponentHostCapability NativeActor impl into runtime.rs under
#[runtime] and host #[actor(singleton)] on the struct (ADR-0123). Refs #2378.
